Feat: relacionamento listaXtarefa

+++ src/main/java/com/squad03/flap/DTO/BuscaLista.java
@@ -2,7 +2,7 @@ package com.squad03.flap.DTO;
 
 import com.squad03.flap.model.Lista;
 
-public record BuscaLista(int id, String nome, int posicao, String cor) {
+public record BuscaLista(Long id, String nome, int posicao, String cor) {
     public BuscaLista (Lista lista){
         this(lista.getId(), lista.getNome(), lista.getPosicao(), lista.getCor());
     }
